+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 * GET requests are now signed if the django configuration includes FEDERATION_USER which is used to fetch that
   user's private key.
 
+* Added Video and Audio objects.
+
+* Process Activitypub reply collections.
+
 ### Fixed
 
 * Signatures are not verified and the corresponding payload is dropped is no public key is found.